django/requirement.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
Pillow==10.4.0
python-decouple==3.8